requirements.txt
streamlit==1.12.0
Pillow==9.3.0
transformers==4.26.0
pdf2image==1.16.0
python-docx==0.8.11
torch==1.13.1
byaldi==1.0.0
fuzzywuzzy==0.18.0
